third_party/blink/web_tests/http/tests/websocket/permessage-deflate-parameter-expected.txt

Test WebSocket permessage-deflate extension.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing query: "client_max_window_bits=8"
PASS ws.extensions.search('permessage-deflate') != -1 is true
PASS ws.extensions.search('client_max_window_bits=8') != -1 is true
PASS event.data is firstMessage
PASS event.data is secondMessage
onclose() was called.
PASS closeEvent.wasClean is true
Testing query: "client_no_context_takeover"
PASS ws.extensions.search('permessage-deflate') != -1 is true
PASS ws.extensions.search('client_no_context_takeover') != -1 is true
PASS event.data is firstMessage
PASS event.data is secondMessage
onclose() was called.
PASS closeEvent.wasClean is true
Testing query: "client_max_window_bits=8&client_no_context_takeover"
PASS ws.extensions.search('permessage-deflate') != -1 is true
PASS ws.extensions.search('client_max_window_bits=8') != -1 is true
PASS ws.extensions.search('client_no_context_takeover') != -1 is true
PASS event.data is firstMessage
PASS event.data is secondMessage
onclose() was called.
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
